+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/Kompendium.kt
@@ -5,11 +5,12 @@ import io.ktor.http.HttpMethod
 import io.ktor.routing.Route
 import io.ktor.routing.method
 import io.ktor.util.pipeline.PipelineInterceptor
-import kotlin.reflect.full.findAnnotation
+import kotlin.reflect.KType
+import kotlin.reflect.typeOf
 import org.leafygreens.kompendium.Kontent.generateKontent
-import org.leafygreens.kompendium.annotations.KompendiumRequest
-import org.leafygreens.kompendium.annotations.KompendiumResponse
 import org.leafygreens.kompendium.models.meta.MethodInfo
+import org.leafygreens.kompendium.models.meta.RequestInfo
+import org.leafygreens.kompendium.models.meta.ResponseInfo
 import org.leafygreens.kompendium.models.oas.OpenApiSpec
 import org.leafygreens.kompendium.models.oas.OpenApiSpecInfo
 import org.leafygreens.kompendium.models.oas.OpenApiSpecMediaType
@@ -18,8 +19,8 @@ import org.leafygreens.kompendium.models.oas.OpenApiSpecPathItemOperation
 import org.leafygreens.kompendium.models.oas.OpenApiSpecReferenceObject
 import org.leafygreens.kompendium.models.oas.OpenApiSpecRequest
 import org.leafygreens.kompendium.models.oas.OpenApiSpecResponse
-import org.leafygreens.kompendium.util.Helpers.COMPONENT_SLUG
 import org.leafygreens.kompendium.util.Helpers.calculatePath
+import org.leafygreens.kompendium.util.Helpers.getReferenceSlug
 
 object Kompendium {
 
@@ -29,94 +30,102 @@ object Kompendium {
     paths = mutableMapOf()
   )
 
+  @OptIn(ExperimentalStdlibApi::class)
   inline fun <reified TParam : Any, reified TResp : Any> Route.notarizedGet(
     info: MethodInfo,
     noinline body: PipelineInterceptor<Unit, ApplicationCall>
-  ): Route = generateComponentSchemas<Unit, TResp>() {
+  ): Route = notarizationPreFlight<Unit, TResp>() { requestType, responseType ->
     val path = calculatePath()
     openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
-    openApiSpec.paths[path]?.get = info.parseMethodInfo<Unit, TResp>()
+    openApiSpec.paths[path]?.get = info.parseMethodInfo(HttpMethod.Get, requestType, responseType)
     return method(HttpMethod.Get) { handle(body) }
   }
 
   inline fun <reified TParam : Any, reified TReq : Any, reified TResp : Any> Route.notarizedPost(
     info: MethodInfo,
     noinline body: PipelineInterceptor<Unit, ApplicationCall>
-  ): Route = generateComponentSchemas<TReq, TResp>() {
+  ): Route = notarizationPreFlight<TReq, TResp>() { requestType, responseType ->
     val path = calculatePath()
     openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
-    openApiSpec.paths[path]?.post = info.parseMethodInfo<TReq, TResp>()
+    openApiSpec.paths[path]?.post = info.parseMethodInfo(HttpMethod.Post, requestType, responseType)
     return method(HttpMethod.Post) { handle(body) }
   }
 
   inline fun <reified TParam : Any, reified TReq : Any, reified TResp : Any> Route.notarizedPut(
     info: MethodInfo,
     noinline body: PipelineInterceptor<Unit, ApplicationCall>,
-  ): Route = generateComponentSchemas<TReq, TResp>() {
+  ): Route = notarizationPreFlight<TReq, TResp>() { requestType, responseType ->
     val path = calculatePath()
     openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
-    openApiSpec.paths[path]?.put = info.parseMethodInfo<TReq, TResp>()
+    openApiSpec.paths[path]?.put = info.parseMethodInfo(HttpMethod.Put, requestType, responseType)
     return method(HttpMethod.Put) { handle(body) }
   }
 
   inline fun <reified TParam : Any, reified TResp : Any> Route.notarizedDelete(
     info: MethodInfo,
     noinline body: PipelineInterceptor<Unit, ApplicationCall>
-  ): Route = generateComponentSchemas<Unit, TResp> {
+  ): Route = notarizationPreFlight<Unit, TResp> { requestType, responseType ->
     val path = calculatePath()
     openApiSpec.paths.getOrPut(path) { OpenApiSpecPathItem() }
-    openApiSpec.paths[path]?.delete = info.parseMethodInfo<Unit, TResp>()
+    openApiSpec.paths[path]?.delete = info.parseMethodInfo(HttpMethod.Delete, requestType, responseType)
     return method(HttpMethod.Delete) { handle(body) }
   }
 
-  inline fun <reified TReq, reified TResp> MethodInfo.parseMethodInfo() = OpenApiSpecPathItemOperation(
+  // TODO here down is a mess, needs refactor once core functionality is in place
+  fun MethodInfo.parseMethodInfo(
+    method: HttpMethod,
+    requestType: KType,
+    responseType: KType
+  ) = OpenApiSpecPathItemOperation(
     summary = this.summary,
     description = this.description,
     tags = this.tags,
     deprecated = this.deprecated,
-    responses = parseResponseAnnotation<TResp>()?.let { mapOf(it) },
-    requestBody = parseRequestAnnotation<TReq>()
+    responses = responseType.toSpec(responseInfo)?.let { mapOf(it) },
+    requestBody = if (method != HttpMethod.Get) requestType.toSpec(requestInfo) else null
   )
 
-  inline fun <reified TReq : Any, reified TResp : Any> generateComponentSchemas(
-    block: () -> Route
+  @OptIn(ExperimentalStdlibApi::class)
+  inline fun <reified TReq : Any, reified TResp : Any> notarizationPreFlight(
+    block: (KType, KType) -> Route
   ): Route {
     val responseKontent = generateKontent<TResp>()
     val requestKontent = generateKontent<TReq>()
     openApiSpec.components.schemas.putAll(responseKontent)
     openApiSpec.components.schemas.putAll(requestKontent)
-    return block.invoke()
+    val requestType = typeOf<TReq>()
+    val responseType = typeOf<TResp>()
+    return block.invoke(requestType, responseType)
   }
 
-  inline fun <reified TReq> parseRequestAnnotation(): OpenApiSpecRequest? = when (TReq::class) {
+  // TODO These two lookin' real similar 👀 Combine?
+  private fun KType.toSpec(requestInfo: RequestInfo?): OpenApiSpecRequest? = when (this) {
     Unit::class -> null
-    else -> when (val anny = TReq::class.findAnnotation<KompendiumRequest>()) {
+    else -> when (requestInfo) {
       null -> null
       else -> OpenApiSpecRequest(
-        description = anny.description,
-        content = anny.mediaTypes.associate {
-          val ref = OpenApiSpecReferenceObject("$COMPONENT_SLUG/${TReq::class.simpleName}")
-          val mediaType = OpenApiSpecMediaType.Referenced(ref)
-          Pair(it, mediaType)
+        description = requestInfo.description,
+        content = requestInfo.mediaTypes.associateWith {
+          val ref = getReferenceSlug()
+          OpenApiSpecMediaType.Referenced(OpenApiSpecReferenceObject(ref))
         }
       )
     }
   }
 
-  inline fun <reified TResp> parseResponseAnnotation(): Pair<Int, OpenApiSpecResponse>? = when (TResp::class) {
-    Unit::class -> null
-    else -> when (val anny = TResp::class.findAnnotation<KompendiumResponse>()) {
-      null -> null
+  private fun KType.toSpec(responseInfo: ResponseInfo?): Pair<Int, OpenApiSpecResponse>? = when (this) {
+    Unit::class -> null // TODO Maybe not though? could be unit but 200 🤔
+    else -> when (responseInfo) {
+      null -> null // TODO again probably revisit this
       else -> {
         val specResponse = OpenApiSpecResponse(
-          description = anny.description,
-          content = anny.mediaTypes.associate {
-            val ref = OpenApiSpecReferenceObject("$COMPONENT_SLUG/${TResp::class.simpleName}")
-            val mediaType = OpenApiSpecMediaType.Referenced(ref)
-            Pair(it, mediaType)
+          description = responseInfo.description,
+          content = responseInfo.mediaTypes.associateWith {
+            val ref = getReferenceSlug()
+            OpenApiSpecMediaType.Referenced(OpenApiSpecReferenceObject(ref))
           }
         )
-        Pair(anny.status, specResponse)
+        Pair(responseInfo.status, specResponse)
       }
     }
   }