+++ kompendium-core/src/main/kotlin/org/leafygreens/kompendium/models/meta/MethodInfo.kt
@@ -1,8 +1,11 @@
 package org.leafygreens.kompendium.models.meta
 
+// TODO Seal and extend by method type?
 data class MethodInfo(
   val summary: String,
   val description: String? = null,
+  val responseInfo: ResponseInfo? = null,
+  val requestInfo: RequestInfo? = null,
   val tags: Set<String> = emptySet(),
   val deprecated: Boolean = false
 )